Fixed some MSVC warnings about pointer truncation
In this case, those warnings are correct however because they made their 'long' 32-bit. Need to replace all instances of the raw type 'long' with use of a stdint.h type

+++ test/crunch/testCrunch.c
@@ -242,7 +242,7 @@ void testAssertGreaterThan()
 {
 	srand(time(NULL));
 	do
-		value = (long)genPtr();
+		value = (long)(uintptr_t)genPtr();
 	while (ptr == NULL);
 	assertGreaterThan(value, 0);
 	tryShouldFail(testAssertGreaterThan1);
@@ -255,7 +255,7 @@ void testAssertLessThan()
 {
 	srand(time(NULL));
 	do
-		value = (long)genPtr();
+		value = (long)(uintptr_t)genPtr();
 	while (ptr == NULL);
 	assertLessThan(0, value);
 	tryShouldFail(testAssertLessThan1);
